move : station 패키지 생성 및 구조 이동
