+++ src/main/java/subway/station/StationResponse.java
@@ -1,8 +1,11 @@
-package subway;
+package subway.station;
 
 public class StationResponse {
     private Long id;
     private String name;
+
+    public StationResponse() {
+    }
 
     public StationResponse(Long id, String name) {
         this.id = id;
